Modernize for QGIS 4 and overhaul the release process
Make the plugin QGIS 4 (Qt6) compatible and replace the outdated
Makefile-based release flow with qgis-plugin-ci + GitHub Actions.

QGIS 4 / Qt6 compatibility:
- Scope Qt enums and use exec() in gbif_webservices.show_warning()
  (QMessageBox.Warning/.Yes/.No and exec_() were removed in Qt6). The
  scoped forms work under both PyQt5 and PyQt6, so QGIS 3 keeps working.
- Drop the Qt5-compiled resources_rc.py (incompatible with Qt6) and load
  the toolbar icon from a filesystem path instead.
- Remove the unused, pyuic-generated qgis_occurrences_dialog_base.py; the
  dialog is built at runtime from the .ui via uic.loadUiType.

Release process:
- Move plugin sources into a GBIFOccurrences/ subdirectory as required by
  qgis-plugin-ci (the folder name is kept to preserve existing installs).
- Add .qgis-plugin-ci, a tag-triggered GitHub Actions release workflow
  (GitHub release + publish to plugins.qgis.org), and RELEASING.md.
- Add a uv-managed tooling environment (pyproject.toml, .python-version,
  uv.lock) and CONTRIBUTING.md documenting local build/test.
- Remove obsolete tooling: Makefile, plugin_upload.py, pylintrc, the
  translation scripts, and the leftover Afrikaans translation scaffolding.

Fixes:
- count_occurrences() no longer raises UnboundLocalError on an unexpected
  GBIF response; it returns 0 for an empty result set and raises a clear
  GBIFApiError otherwise. Remove leftover debug prints.
- Remove accidentally-committed local state (user-history.db,
  symbology-style.db).
