--- resources.qrc
@@ -1,5 +1,0 @@
-<RCC>
-    <qresource prefix="/plugins/GBIFOccurrences" >
-        <file>icon.png</file>
-    </qresource>
-</RCC>
