--- resources_rc.py
@@ -1,86 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Resource object code
-#
-# Created by: The Resource Compiler for PyQt5 (Qt v5.9.4)
-#
-# WARNING! All changes made in this file will be lost!
-
-from qgis.PyQt import QtCore
-
-qt_resource_data = b"\
-\x00\x00\x01\x61\
-\x89\
-\x50\x4e\x47\x0d\x0a\x1a\x0a\x00\x00\x00\x0d\x49\x48\x44\x52\x00\
-\x00\x00\x18\x00\x00\x00\x17\x08\x03\x00\x00\x00\x26\xff\x7f\x1f\
-\x00\x00\x00\x90\x50\x4c\x54\x45\x80\x80\x80\xc0\xc0\xc0\x40\x40\
-\x40\x73\xbd\x5f\xa2\xd3\x94\xf4\xfa\xf2\xd1\xe9\xca\x50\xad\x36\
-\xb9\xde\xaf\x67\xb8\x51\x8a\xc8\x7a\xe8\xf4\xe5\x5b\xb2\x44\xdd\
-\xef\xd8\xc5\xe4\xbd\x96\xce\x87\xae\xd9\xa2\xd1\xe5\xb9\x7f\xc3\
-\x6c\xba\xd7\x95\xfa\xfc\xf7\xe8\xf2\xdc\xa9\xcd\x7b\x20\x20\x20\
-\xc6\xde\xa7\xf0\xf0\xf0\xaf\xd1\x84\x50\x50\x50\xe0\xe0\xe0\xa0\
-\xa0\xa0\xcb\xe1\xb0\xb4\xd4\x8d\x10\x10\x10\xee\xf5\xe5\x90\x90\
-\x90\xf4\xf9\xee\xd7\xe8\xc2\x70\x70\x70\x60\x60\x60\xd0\xd0\xd0\
-\xe2\xef\xd3\xdd\xeb\xca\xb0\xb0\xb0\x30\x30\x30\xc0\xdb\x9e\xa3\
-\xca\x72\x00\x00\x00\x44\xa7\x29\x01\x9e\x79\x57\x00\x00\x00\x01\
-\x74\x52\x4e\x53\x00\x40\xe6\xd8\x66\x00\x00\x00\x01\x62\x4b\x47\
-\x44\x00\x88\x05\x1d\x48\x00\x00\x00\x09\x70\x48\x59\x73\x00\x00\
-\x0b\x13\x00\x00\x0b\x13\x01\x00\x9a\x9c\x18\x00\x00\x00\x07\x74\
-\x49\x4d\x45\x07\xde\x0c\x02\x0f\x20\x16\xe1\x15\x24\xbb\x00\x00\
-\x00\x4a\x49\x44\x41\x54\x28\xcf\x63\x60\x80\x00\x7d\x06\x1c\x00\
-\x97\x84\x3e\x6e\x09\x7d\x32\x25\x74\x81\x00\x9b\x04\x86\x38\x54\
-\x42\x17\x9f\x04\x03\x0e\x09\x8a\x9d\xab\x0f\x06\xd8\x45\xa1\x40\
-\x08\x35\x98\x50\xe4\x50\x83\x95\x17\x8b\x14\x16\x13\xa9\x2a\xc3\
-\x40\xac\x0c\x03\x4e\x09\x06\x06\x66\x88\x04\x00\xbf\x23\x1f\x60\
-\x8e\x01\x2a\xe1\x00\x00\x00\x00\x49\x45\x4e\x44\xae\x42\x60\x82\
-\
-"
-
-qt_resource_name = b"\
-\x00\x07\
-\x07\x3b\xe0\xb3\
-\x00\x70\
-\x00\x6c\x00\x75\x00\x67\x00\x69\x00\x6e\x00\x73\
-\x00\x0f\
-\x02\xe2\x51\xe3\
-\x00\x47\
-\x00\x42\x00\x49\x00\x46\x00\x4f\x00\x63\x00\x63\x00\x75\x00\x72\x00\x72\x00\x65\x00\x6e\x00\x63\x00\x65\x00\x73\
-\x00\x08\
-\x0a\x61\x5a\xa7\
-\x00\x69\
-\x00\x63\x00\x6f\x00\x6e\x00\x2e\x00\x70\x00\x6e\x00\x67\
-"
-
-qt_resource_struct_v1 = b"\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x14\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x38\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-"
-
-qt_resource_struct_v2 = b"\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x14\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x38\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x65\xfc\x72\xf1\x60\
-"
-
-qt_version = QtCore.qVersion().split('.')
-if qt_version < ['5', '8', '0']:
-    rcc_version = 1
-    qt_resource_struct = qt_resource_struct_v1
-else:
-    rcc_version = 2
-    qt_resource_struct = qt_resource_struct_v2
-
-def qInitResources():
-    QtCore.qRegisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-def qCleanupResources():
-    QtCore.qUnregisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-qInitResources()
